requirements.txt
Flask==2.3.2
Werkzeug==2.3.6
Jinja2==3.1.2
ultralytics>=8.0.196
opencv-python==4.8.0.74
Pillow==10.0.0
numpy==1.24.3
torch>=1.7.0,<2.6.0
torchvision>=0.8.0,<0.20.0
matplotlib==3.7.1
seaborn==0.12.2
pandas==2.0.3
requests==2.31.0
tqdm==4.65.0
PyYAML==6.0
psutil==5.9.5 